+++ examples/react/app/src/components/copyfactory/copyTrade.tsx
@@ -87,7 +87,7 @@ export function CopyfactoryCopyTrade({
         );
       }
       
-      const strategies = await configurationApi.getStrategies();
+      const strategies = await configurationApi.getStrategiesWithInfiniteScrollPagination();
       const strategy = strategies.find((s: any) => s.accountId === providerMetaapiAccount.id);
       
       const strategyId = !strategy