+++ examples/react/app/src/components/copyfactory/telegram.tsx
@@ -103,7 +103,7 @@ export function CopyfactoryTelegram({
 
         log('Provider account received');
       
-        const strategies = await configurationApi.getStrategies();
+        const strategies = await configurationApi.getStrategiesWithInfiniteScrollPagination();
         const strategy = strategies.find((s) => s.accountId === providerMetaapiAccount.id);
         
         const strategyId = !strategy