+++ examples/vue/app/src/components/copyfactory/copyTrade.vue
@@ -68,7 +68,7 @@ const makeRequest = async () => {
       );
     }
     
-    const strategies = await configurationApi.getStrategies();
+    const strategies = await configurationApi.getStrategiesWithInfiniteScrollPagination();
     const strategy = strategies.find((s) => s.accountId === providerMetaapiAccount.id);
     
     const strategyId = !strategy