+++ examples/vue/app/src/components/copyfactory/externalSignal.vue
@@ -67,7 +67,7 @@ const makeRequest = async () => {
     
     const configurationApi = copyFactory.configurationApi;
     
-    const strategies = await configurationApi.getStrategies();
+    const strategies = await configurationApi.getStrategiesWithInfiniteScrollPagination();
     const strategy = strategies.find((s) => s.accountId === providerMetaapiAccount.id);
     
     const strategyId = !strategy